+++ jeditor/core/scenemanager.py
@@ -1,5 +1,6 @@
+from jeditor.core.nodefactory import JNodeFactory
 from jeditor.core.graphicedge import JGraphicEdge
-from typing import List
+from typing import Dict, List
 
 from PyQt5.QtCore import QPointF
 from .contentwidget import JNodeContent
@@ -11,56 +12,30 @@ from .constants import (
     GRSOCKET_TYPE_INPUT,
     GRSOCKET_TYPE_OUTPUT,
 )
-from .graphicscene import JiglsGraphicScene
-
-
-def DevNode(inputMulti, outputMulti, inputs=1, output=1):
-    node = JGraphicNode(nodeContent=JNodeContent())
-    for _ in range(inputs):
-        node.socketManager.AddSocket(
-            type=GRSOCKET_TYPE_INPUT, multiConnection=inputMulti
-        )
-    for _ in range(output):
-        node.socketManager.AddSocket(
-            type=GRSOCKET_TYPE_OUTPUT, multiConnection=outputMulti
-        )
-    return node
+from .graphicscene import JGraphicScene
+from pprint import pprint
+import json
 
 
 class JSceneManager:
     def __init__(self) -> None:
 
-        self.nodes: List[JGraphicNode] = []
-        self.edges = []
-
         self.initUI()
         self._debug()
-        # node.title = "david"
 
     def initUI(self):
-        self._graphicsScene = JiglsGraphicScene()
+        self._graphicsScene = JGraphicScene()
+        self._nodeFactory = JNodeFactory()
         self._graphicsScene.SetGraphicsSceneWH(GRSCENE_WIDTH, GRSCENE_HEIGHT)
 
     @property
     def graphicsScene(self):
         return self._graphicsScene
 
-    def AddNode(self, node: JGraphicNode):
-        self.nodes.append(node)
-
-    def AddEdge(self, edge):
-        self.edges.append(edge)
-
-    def RemoveNode(self, node: JGraphicNode):
-        self.nodes.remove(node)
-
-    def RemoveEdge(self, edge):
-        self.nodes.remove(edge)
-
     def _debug(self):
-        node1 = DevNode(False, False)
-        node2 = DevNode(True, True)
-        node3 = DevNode(True, False)
+        node1 = self._nodeFactory.CreateNode(None, False, False)
+        node2 = self._nodeFactory.CreateNode(None, True, True)
+        node3 = self._nodeFactory.CreateNode(None, True, False)
         # node4 = JGraphicNode(inSockets=1, outSockets=1, nodeContent=JNodeContent())
 
         node1.setPos(QPointF(-350, -250))
@@ -70,21 +45,63 @@ class JSceneManager:
         self.graphicsScene.addItem(node1)
         self.graphicsScene.addItem(node2)
         self.graphicsScene.addItem(node3)
-        # self.graphicsScene.addItem(node4)
 
-        # edge1 = JGraphicEdge(
-        #     node2.socketManager.GetOutputSocketByIndex(0),
-        #     node3.socketManager.GetInputSocketByIndex(0),
-        # )
-        # edge2 = JGraphicEdge(
-        #     node2.socketManager.GetOutputSocketByIndex(0),
-        #     node1.socketManager.GetInputSocketByIndex(0),
-        #     edgePathType=GREDGE_PATH_BEZIER,
-        # )
+    def Serialize(self) -> Dict:
+        node: List[Dict] = []
+        edge: List[Dict] = []
+        for item in self._graphicsScene.items():
+            if isinstance(item, JGraphicNode):
+                node.append(item.Serialize())
+            if isinstance(item, JGraphicEdge):
+                edge.append(item.Serialize())
+        return {"nodes": node, "edges": edge}
 
-        # self.graphicsScene.addItem(edge1)
-        # self.graphicsScene.addItem(edge2)
+    def Deserialize(self, data: Dict):
+        self._graphicsScene.clear()
+        for node in data["nodes"]:
+            instanceNode = JGraphicNode.Deserialize(node)
+            self._graphicsScene.addItem(instanceNode)
 
-        # node3 = JGraphicNode(inSockets=2, outSockets=2, nodeContent=JNodeContent())
-        # node3.setPos(QPointF(200, -150))
-        # self.graphicsScene.addItem(node3)
+        for edge in data["edges"]:
+            nodes = list(
+                filter(
+                    lambda x: x.nodeIdentifier  # type:ignore
+                    in [edge["sourceNodePID"], edge["destinationNodePID"]],
+                    list(
+                        filter(
+                            lambda item: isinstance(item, JGraphicNode),
+                            self._graphicsScene.items(),
+                        )
+                    ),
+                )
+            )
+            assert len(nodes) == 2
+            startSocket = None
+            destinationSocket = None
+            for node in nodes:
+                assert isinstance(node, JGraphicNode)
+                if node.nodeIdentifier == edge["sourceNodePID"]:
+                    startSocket = node.socketManager.GetSocketByIndex(
+                        edge["sourceNodeIndex"]
+                    )
+                if node.nodeIdentifier == edge["destinationNodePID"]:
+                    destinationSocket = node.socketManager.GetSocketByIndex(
+                        edge["destinationNodeIndex"]
+                    )
+
+            self._graphicsScene.addItem(
+                JGraphicEdge.Deserialize(
+                    identifier=edge["identifier"],
+                    startSocket=startSocket,
+                    destinationSocket=destinationSocket,
+                )
+            )
+
+    def SaveToFile(self):
+        with open("graph.json", "w") as file:
+            json.dump(obj=self.Serialize(), fp=file)
+
+    def LoadFromFile(self) -> Dict:
+        with open("graph.json", "r") as file:
+            data = json.load(file)
+            return data